[AAE-4608] Use smartrunner version 2 and add SuperCache (workspaces + S3 combo)! (#6704)
* [AAE-4608] Use smartrunner version 2

* Fix npm installing and do not depend on develop cache

* Add Travis workspaces the simplest way

* Add missing dependencies

* Remove unnecessary npm install on lint

* Fix travis indentation

* Fix e2e missing adf packages

* Trying to fix Travis

* Remove debug information

* Use Travis workspaces for smartrunner and built artefacts caching instead of S3

* Implement nuclear caching for node_modules

* Quiet snyk running, since the log length kills Travis

* Maybe with -q

* Quieting snyk

* Why displaying the added packages??????

* Fix S3 persister/retriever scripts

* Restore back parallel build and lint, to gain 10 minutes

+++ scripts/ci/utils/artifact-from-s3.sh
@@ -23,9 +23,15 @@ then
 fi
 
 test ! -d $OUTPUT && mkdir -p $OUTPUT
-aws s3 cp $ARTIFACT ./s3-artifact.tmp
-echo 'artifact download done'
-tar -xvf ./s3-artifact.tmp -C $OUTPUT >&/dev/null
-echo 'tar the artifact done'
-rm ./s3-artifact.tmp
-echo 'remove tmp file'
+
+IS_PRESENT="$(aws s3 ls $ARTIFACT | wc -l | tr -d ' ')"
+if [ "${IS_PRESENT}" == "1" ]
+then
+  echo "File ${ARTIFACT} is present. Copying"
+  aws s3 cp $ARTIFACT ./s3-artifact.tmp
+  tar -xf ./s3-artifact.tmp -C $OUTPUT
+  rm ./s3-artifact.tmp
+else
+  echo "File ${ARTIFACT} not present"
+  exit 1;
+fi
